requirements.txt
torch==2.0.0
torchvision==0.15.0
numpy==1.24.3
matplotlib==3.7.1
Pillow==10.0.0
scikit-image==0.21.0
opencv-python==4.8.0.74
tensorboard==2.13.0
tqdm==4.65.0
scipy==1.11.1
jupyter==1.0.0
ipython==8.14.0
pyyaml==6.0
jsonschema==4.18.0
albumentations==1.3.0
torchsummary==1.5.1
segmentation-models-pytorch==0.3.3
